farm-node/requirements.txt
RPi.GPIO==0.7.1
requests==2.31.0
